Quota: rather than using the permission's "owner" in the quota system (control of which is delegated to users), track the domain which *created* a node
The 'creator' of a node is immutable and tracked by store. It is not
possible to change the creator of a node. Therefore if a domain creates
a node, it is accounted to that domain's quota, even if it sets the
permissions owner to someone else.

It is not possible for one domain to subtract from the quota of another
domain.