CA-100569: Perform a 3-way merge a the quota after a transaction.
At a beginning of a transaction, the quotas from the global store are duplicated
and modified by the transaction. If during the transaction, an action associated
to no transaction were concurrently executed, the quotas of the global store
were updated, but the change from the transaction would be unaccounted in the
quota.
This commit fix this problem by keeping another copy of the quota at the
beginning of the transaction, and performing a 3-way merge between the quotas
from the transaction and the "original" copy of the quota onto the quotas
of the global store.

Signed-off-by: Jerome Maloberti <jerome.maloberti@citrix.com>