server: persist quota limits in the database
The global default settings are stored in persistent reference
cells (module PRef) while the per-domain overrides are stored
in persistent maps (module PMap).

The actual limits in force for a particular domain are computed
in the Lwt-based packet processing loop and passed down to the
pure Store/Transaction/Node modules.

This patch removes the ability to add a per-domain override for
the maximum length of watch event queues. Implementing this is
awkward because domid x generates the watch event, but the limits
for domid y (where x may not be equal to y) need to apply.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>