Fix segfault in mmap stub
The C code was returning an Io_page.t option, while the declared
interface was a raw Io_page.t. This function is only needed to map
dom0's ring, and if we can't do that we might as well abort(). So
change the C code to return a raw Io_page.t.

Signed-off-by: David Scott <dave.scott@citrix.com>