Remove client_unix/client_test.ml (now elsewhere)
This source file has not been compiled in this repository since
Anil's change in July 2013 to fix a circular dependency:
"Fix build cycle between xenstore and xenstore_transport"

Recently I copied the file from here to the ocaml-xenstore-clients
repository, and it is being built there now. Therefore it can be
deleted from here.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>