Terminating replacements for transaction function
The Xs_client_unix.transaction function would retry for ever if some
client were to keep making writes conflicting with the transaction.

This commit replaces `transaction` with `transaction_one_try` and
`transaction_attempts` which will make only a limited number of
attempts before abandoning the transaction on conflict and passing
the Eagain exception up to the caller.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>