server: disable the tests that are currently broken
The server library (`xenstore.server`) is not used anywhere and it's
in a semi-broken state.

By constrast the main library (`xenstore`) is critical and in heavy
use.

This patches makes `make test` run all the critical main library tests
and only the currently-working subset of the unused server library tests.

Signed-off-by: David Scott <dave@recoil.org>