Fix crash if watch quota is exceeded
If watching a path fails (e.g. with `E2BIG` because we exceeded the
quota) then before reporting the error we try to unwatch all the paths,
even the ones we didn't manage to watch. This then fails with a
different error.

Now we only unwatch paths that we managed to watch.